+++ include/linux/sched.h
@@ -1794,12 +1794,17 @@ extern int task_free_unregister(struct notifier_block *n);
 extern int sched_set_window(u64 window_start, unsigned int window_size);
 extern unsigned long sched_get_busy(int cpu);
 extern void sched_set_io_is_busy(int val);
+extern unsigned long sched_cpu_util(int cpu);
 #else
 static inline int sched_set_window(u64 window_start, unsigned int window_size)
 {
 	return -EINVAL;
 }
 static inline unsigned long sched_get_busy(int cpu)
+{
+	return 0;
+}
+static inline unsigned long sched_cpu_util(int cpu)
 {
 	return 0;
 }